Change User info page style and function
add user evaluation page for viewing what other user or me evaluate and
the function that evaluate destination on that page.

+++ tourofall/src/main/java/net/bulldozer/tourofall/user/service/UserService.java
@@ -18,9 +18,9 @@ public interface UserService {
 	public Collection<AnswerRenderingModel> getAnswersByUserId(long id);
 	public Collection<ReviewRenderingModel> getReviewsByUserId(long id);
 	public Collection<EvaluationRenderingModel> getEvaluationsByUserId(long id);
-	public int getQuestionsSizeByUserId(long id);
-	public int getAnswersSizeByUserId(long id);
-	public int getReviewsSizeByUserId(long id);
-	public int getEvaluationsSizeByUserId(long id);
+	public long getQuestionsSizeByUserId(long id);
+	public long getAnswersSizeByUserId(long id);
+	public long getReviewsSizeByUserId(long id);
+	public long getEvaluationsSizeByUserId(long id);
 	public UserModificationForm getUserModificationFormByUserId(long id);
 }
